Fix responsiveness of RGA GUI disassembly view
The GUI becomes unresponsive for several seconds whenever the disassembly view
gains or loses focus due to the code unnecessarily applying a stylesheet to
the entire frame just to change border color, which seems to be a leftover
functionality, considering that the color was only affected by the API used
which is now chosen at application launch so the border color does not provide
much value. According to the Qt documentation applications should not call
setStylesheet() outside of the constructor because it's very costly and the
current code calls it on focus change.

This change fixes that by getting rid of all the code using setStylesheet()
to change the disassembly view's border.

Also, some missing header includes were added which most likely were
uncovered by compiling against Qt 5.11.2 contrarily to Qt 5.9.2 recommended
in the RGA documentation.